feat: set logging callback, not working yet
